+++ lib/PuppeteerSharp.Tests/FrameTests/GoToTests.cs
@@ -1,5 +1,7 @@
-﻿using System.Net;
+﻿using System.Collections.Generic;
+using System.Net;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -20,6 +22,7 @@ namespace PuppeteerSharp.Tests.FrameTests
             Assert.Contains("/frames/frame.html", Page.Frames[1].Url);
             var response = await Page.Frames[1].GoToAsync(TestConstants.EmptyPage);
             Assert.Equal(HttpStatusCode.OK, response.Status);
+            Assert.Same(response.Frame, Page.Frames[1]);
         }
 
         [Fact]
@@ -33,6 +36,48 @@ namespace PuppeteerSharp.Tests.FrameTests
             await Page.QuerySelectorAsync("iframe").EvaluateFunctionAsync("frame => frame.remove()");
             var exception = await Assert.ThrowsAsync<NavigationException>(async () => await navigationTask);
             Assert.Equal("Navigating frame was detached", exception.Message);
-        } 
+        }
+
+        [Fact]
+        public async Task ShouldReturnMatchingResponses()
+        {
+            // Disable cache: otherwise, chromium will cache similar requests.
+            await Page.SetCacheEnabledAsync(false);
+            await Page.GoToAsync(TestConstants.EmptyPage);
+            // Attach three frames.
+            var frameTasks = new List<Task<Frame>>
+            {
+                FrameUtils.AttachFrameAsync(Page, "frame1", TestConstants.EmptyPage),
+                FrameUtils.AttachFrameAsync(Page, "frame2", TestConstants.EmptyPage),
+                FrameUtils.AttachFrameAsync(Page, "frame3", TestConstants.EmptyPage)
+            };
+            await Task.WhenAll(frameTasks);
+
+            // Navigate all frames to the same URL.
+            var serverResponses = new List<TaskCompletionSource<string>>();
+            Server.SetRoute("/one-style.html", async (context) =>
+            {
+                var tcs = new TaskCompletionSource<string>();
+                serverResponses.Add(tcs);
+                await context.Response.WriteAsync(await tcs.Task);
+            });
+
+            var navigations = new List<Task<Response>>();
+            for (var i = 0; i < 3; ++i)
+            {
+                var waitRequestTask = Server.WaitForRequest("/one-style.html");
+                navigations.Add(frameTasks[i].Result.GoToAsync(TestConstants.ServerUrl + "/one-style.html"));
+                await waitRequestTask;
+            }
+            // Respond from server out-of-order.
+            var serverResponseTexts = new string[] { "AAA", "BBB", "CCC" };
+            for (var i = 0; i < 3; ++i)
+            {
+                serverResponses[i].TrySetResult(serverResponseTexts[i]);
+                var response = await navigations[i];
+                Assert.Same(frameTasks[i].Result, response.Frame);
+                Assert.Equal(serverResponseTexts[i], await response.TextAsync());
+            }
+        }
     }
 }
